+++ src/i18n/nl.json
@@ -1,32 +1,27 @@
 {
- "back_to_home": "Terug naar huis",
- "close": "Sluit",
- "closed": "Gesloten",
- "hello": "Hallo",
- "label.bio": "Bio",
- "label.celery": "Selderij",
- "label.crustacens": "Schaaldieren",
- "label.eggs": "Eieren",
- "label.fish": "Vis",
- "label.local": "Lokaal",
- "label.low-carb": "Low-Carb",
- "label.lupin": "Lupin",
- "label.milk": "Melk",
- "label.molluscs": "Weekdieren",
- "label.mustard": "Mosterd",
- "label.organic": "Organic",
- "label.peanut": "Pinda",
- "label.raw": "Rauw",
- "label.sesame": "Sesam",
- "label.soy": "Soja",
- "label.sulphurdioxide": "Zwaveldioxide",
- "label.treenut": "Boom noot",
- "label.vegan": "Veganistisch",
- "label.vegetarian": "Vegetarisch",
- "label.wheat": "Tarwe",
- "menu": "Menu",
- "next": "Volgende",
- "or_not_saved": "of niet opgeslagen",
- "page_not_found": "Pagina niet gevonden",
- "previous": "Vorig"
+  "home": "Home",
+  "contact": "Contact",
+  "back_to_home": "Terug naar home",
+  "page_not_found": "Pagina niet gevonden",
+  "or_not_saved": "of niet opgeslagen",
+  "hello": "Hallo",
+  "closed": "Gesloten",
+  "close": "Sluiten",
+  "menu": "Menu",
+  "next": "Volgende",
+  "previous": "Vorige",
+  "select_a_topic": "Selecteer een onderwerp",
+  "email": "E-mail*",
+  "name": "Naam*",
+  "phone": "Telefoon",
+  "message": "Bericht*",
+  "send": "Versturen",
+  "submit": "Verzenden",
+  "contact_missing_fields": "Vul alle verplichte velden in",
+  "contact_error": "Er is een fout opgetreden bij het verzenden van het bericht. Probeer het later opnieuw.",
+  "missing_email": "Voer uw e-mailadres in",
+  "newsletter_error": "Er is een fout opgetreden bij het aanmelden voor de nieuwsbrief. Probeer het later opnieuw.",
+  "subscribe_to_newsletter": "Schrijf u in voor onze nieuwsbrief",
+  "newsletter_already_subscribed": "U bent al ingeschreven op onze nieuwsbrief",
+  "subscribe": "Inschrijven"
 }
